+++ nbautoexport/__init__.py
@@ -1,4 +1,8 @@
 from nbautoexport.export import post_save
-from nbautoexport.utils import __version__
+from nbautoexport.utils import __version__, get_logger
 
-__all__ = [post_save, __version__]
+__all__ = [
+    "post_save",
+    "get_logger",
+]
+__version__