+++ nbautoexport/export.py
@@ -11,7 +11,9 @@ from nbautoexport.sentinel import (
     NbAutoexportConfig,
     SAVE_PROGRESS_INDICATOR_FILE,
 )
-from nbautoexport.utils import cleared_argv
+from nbautoexport.utils import cleared_argv, get_logger
+
+logger = get_logger()
 
 
 class CopyToSubfolderPostProcessor(PostProcessorBase):
@@ -68,21 +70,31 @@ def post_save(model: dict, os_path: str, contents_manager: FileContentsManager):
         os_path (str): the filesystem path to the file just written
         contents_manager (FileContentsManager): FileContentsManager instance that hook is bound to
     """
-    # only do this for notebooks
-    if model["type"] != "notebook":
-        return
+    logger.debug("nbautoexport | Executing nbautoexport.export.post_save ...")
+    try:
+        # only do this for notebooks
+        if model["type"] != "notebook":
+            logger.debug(f"nbautoexport | {os_path} is not a notebook. Nothing to do.")
+            return
 
-    # only do this if we've added the special indicator file to the working directory
-    os_path = Path(os_path)
-    cwd = os_path.parent
-    save_progress_indicator = cwd / SAVE_PROGRESS_INDICATOR_FILE
-    should_convert = save_progress_indicator.exists()
+        # only do this if we've added the special indicator file to the working directory
+        notebook_path = Path(os_path)
+        cwd = notebook_path.parent
+        save_progress_indicator = cwd / SAVE_PROGRESS_INDICATOR_FILE
+        should_convert = save_progress_indicator.exists()
 
-    if should_convert:
-        config = NbAutoexportConfig.parse_file(
-            path=save_progress_indicator, content_type="application/json", encoding="utf-8"
-        )
-        export_notebook(os_path, config=config)
+        if should_convert:
+            logger.info(f"nbautoexport | {save_progress_indicator} found. Exporting notebook ...")
+            config = NbAutoexportConfig.parse_file(
+                path=save_progress_indicator, content_type="application/json", encoding="utf-8"
+            )
+            export_notebook(notebook_path, config=config)
+
+        else:
+            logger.debug(f"nbautoexport | {save_progress_indicator} not found. Nothing to do.")
+        logger.debug("nbautoexport | post_save successful.")
+    except Exception as e:
+        logger.error(f"nbautoexport | post_save failed due to {type(e).__name__}: {e}")
 
 
 def export_notebook(notebook_path: Path, config: NbAutoexportConfig):
@@ -92,8 +104,12 @@ def export_notebook(notebook_path: Path, config: NbAutoexportConfig):
         notebook_path (Path): path to notebook to export with nbconvert
         config (NbAutoexportConfig): configuration
     """
+    logger.info(f"nbautoexport | Exporting {notebook_path} ...")
+    logger.debug(f"nbautoexport | Using export configuration:\n{config.json(indent=2)}")
     with cleared_argv():
         converter = NbConvertApp()
+        converter.log.handlers = logger.handlers
+        converter.log.setLevel(logger.level)
 
         for export_format in config.export_formats:
             if config.organize_by == "notebook":