+++ tests/test_post_save.py
@@ -1,3 +1,5 @@
+import json
+import logging
 import shutil
 
 import nbformat
@@ -12,6 +14,7 @@ from nbautoexport.sentinel import (
     OrganizeBy,
     SAVE_PROGRESS_INDICATOR_FILE,
 )
+from tests.utils import caplog_contains
 
 
 @pytest.fixture()
@@ -38,8 +41,9 @@ def file_contents_manager(notebook_file):
     return config.FileContentsManager
 
 
-def test_post_save(file_contents_manager, notebook_file, notebook_model):
+def test_post_save(file_contents_manager, notebook_file, notebook_model, caplog):
     """Test that post_save function works when FileContentsManager saves based on config file."""
+    caplog.set_level(logging.DEBUG)
 
     config = NbAutoexportConfig(
         export_formats=[ExportFormat.script], organize_by=OrganizeBy.extension
@@ -48,6 +52,13 @@ def test_post_save(file_contents_manager, notebook_file, notebook_model):
         fp.write(config.json())
 
     file_contents_manager.save(notebook_model, path=notebook_file.name)
+
+    assert caplog_contains(
+        caplog,
+        level=logging.INFO,
+        in_msg=f"nbautoexport | Exporting {notebook_file}",
+    )
+    caplog.clear()
 
     assert (notebook_file.parent / "script" / f"{notebook_file.stem}.py").exists()
     assert not (notebook_file.parent / notebook_file.stem / f"{notebook_file.stem}.html").exists()
@@ -64,6 +75,13 @@ def test_post_save(file_contents_manager, notebook_file, notebook_model):
         fp.write(config.json())
 
     file_contents_manager.save(notebook_model, path=notebook_file.name)
+
+    assert caplog_contains(
+        caplog,
+        level=logging.INFO,
+        in_msg=f"nbautoexport | Exporting {notebook_file}",
+    )
+    caplog.clear()
 
     assert not (notebook_file.parent / "script" / f"{notebook_file.stem}.py").exists()
     assert (notebook_file.parent / notebook_file.stem / f"{notebook_file.stem}.html").exists()
@@ -102,3 +120,24 @@ def test_no_config(file_contents_manager, notebook_file, notebook_model):
         notebook_file.parent / ".ipynb_checkpoints",
         notebook_file,
     }
+
+
+def test_invalid_config(file_contents_manager, notebook_file, notebook_model, caplog):
+    """Test that post_save function gracefully logs errors."""
+    config = NbAutoexportConfig(
+        export_formats=[ExportFormat.script], organize_by=OrganizeBy.extension
+    )
+    invalid_config = json.loads(config.json())
+    invalid_config["export_formats"] = ["triplicate"]
+    with (notebook_file.parent / SAVE_PROGRESS_INDICATOR_FILE).open("w", encoding="utf-8") as fp:
+        json.dump(invalid_config, fp)
+
+    # Runs through, since error is caught
+    file_contents_manager.save(notebook_model, path=notebook_file.name)
+
+    assert caplog_contains(
+        caplog,
+        level=logging.ERROR,
+        in_msg="nbautoexport | post_save failed due to ValidationError",
+    )
+    assert not (notebook_file.parent / "script" / f"{notebook_file.stem}.py").exists()